Update footer layout and style
Adjusted the placement and styling of the footer link in layout.tsx. The link now includes additional text 'Made with ❤️ by Jakob' and has changes in css class properties to improve the visual layout and accessibility on larger screens. The change is intended to provide a warmer and a more personal touch to the user interface.