Refactor footer link inclusion in layout.tsx
Changes were made to the way the footer link was wrapped around elements in the layout page. Instead of enclosing every element inside a single Link, individual elements are now linked accordingly. This makes the code cleaner and increases the interpretability of the layout for future modifications or additions. The name of the author is now hyperlinked to their website, allowing for direct navigation.