Handle potential null user data in post fetching
Modified the components and data fetching functions related to posts to handle potentially null user data. The Post network model has been updated to accept null values. Now, when fetching post data, if an associated user's data isn't found, the code will not break but simply return null for that post. This was necessary to prevent application crashes when user data is not available for any post due to various reasons, such as user data deletion. Implemented a try-catch block in the fetchAndFormatUser function to handle clerkClient errors as well.