Add search functionality and enhance post-view
This update introduces a search functionality in the application. It allows users to search for posts using a search query. In addition, the `post-view` component has been enhanced to allow for a custom message to be displayed when there are no posts available. The search functionality is implemented in a new `SearchPage` component and integrated in the existing search form. Detailed validation of the search query is also added for more robustness.